Changed the property hoursWorked under WorkHoursTracker from an int to a double in order to display fractional hours worked. Finally have all the issues worked out with displaying the work history for Employees. Had to do a new migration and updated database

+++ CandyShop/Models/WorkHoursTracker.cs
@@ -19,6 +19,6 @@ namespace CandyShop.Models
         public DateTime breakEnd { get; set; }
         public DateTime clockIn { get; set; }
         public DateTime clockOut { get; set; }        
-        public int hoursWorked { get; set; } //Calculate for each work period, minus break times.
+        public double hoursWorked { get; set; } //Calculate for each work period, minus break times.
     }
 }
